requirements.txt
sounddevice
numpy
scipy
moviepy
requests
groq
python-dotenv
tk
